Added and updated lots of unit tests for the rmgpy.statmech module.
Some tests are new, some are not, and others are fixed (or caused fixes
to the statmech module such that they pass). All of the unit tests now
pass in both pure Python and compiled Cython mode, which was not true
earlier today.

Note that many of the direct tests of the heat capacity, enthalpy,
entropy, density of states, and partition function methods do not do any
output validation, as I haven't been able to come up with a way to do
this that isn't just copying the method to the unit test. Nonetheless,
checking that the method runs successfully is still valuable, though not
as valuable as it could be.