Updated and added lots of unit tests for the rmgpy.graph module.
There were already unit tests for the more complicated methods,
particularly the isomorphism methods. This commit adds a bunch more for
the more straightforward methods, plus a few for the cycle methods. All
of the unit tests pass in both pure Python and compiled Cython mode.

Note that the Vertex and Edge classes can really only be properly tested
when part of a graph; thus, these unit tests are lumped with those for
the Graph class.