Added unit test for rmgpy.atomtype module.
This unit test ensures that an AtomType object can be pickled and
unpickled with no loss of information. Note that an AtomType cannot
currently be reconstructed from its repr() output, so no unit test
exists for this purpose. Testing the getAtomType() method requires a
complete molecule, so we will defer that unit test for the moment.