Moved rmgpy.chem.molecule module to rmgpy.molecule.
Although this module arguably belongs to "chem", I am intending to
delete the chem subpackage, so everything in it needs to be moved
elsewhere. The functionality in this module is fairly standalone and
used in lots of places, so it makes sense to place it in the top-level
rmgpy package.