Significant refactoring of rmgpy.group module.
The AtomPattern, BondPattern, and MoleculePattern classes were renamed
to GroupAtom, GroupBond, and Group, respectively, to better reflect the
terminology by which we refer to them anyway. (We tend to call them
functional groups rather than substructure patterns.) I've also changed
all (or most) of the string formatting to use Python 3 conventions.
Finally, I've removed the AtomType functionality that was previously
moved to the rmgpy.atomtype module.