Added lots of unit tests for the rmgpy.group module.
As usual, the objective was to get a unit test for as many methods of
each class as possible. Each class has its own unit test class for this
purpose. This test-driven development continues to show its value, as I
managed to find a few minor but potentially sigificant (and difficult to
trace) bugs with these unit tests.