Added and refactored lots of unit tests for rmgpy.molecule module.
Many of the unit tests, especially those focusing on the Atom and Bond
classes, are new. Many others, especially those in the
TestMoleculeSymmetry unit test class, are not new, but were split so
that each unit test tests a single function for a single molecule. The
combination of both of these means that the number of unit tests for
this module alone is very high, and that's no surprise, as there's a lot
of functionality here to test.

Note that this is the first set of unit tests committed where some unit
tests fail. These are unit tests that ought to pass someday; for now the
tests are simply reminders of things to address in the future. (Aside:
estimating symmetry numbers from 2D connectivity data is hard!)