Fixed bug in atom type declaration for H.
The bug was that R!H was declared as a generic for H, when clearly H is
not in the set of nodes that R!H can represent. (The reverse was defined
correctly.) It is not known just how much of a bug this was, but it's
nonetheless important to get right. Unit tests should help reduce the
likelihood of this issue occurring in the future.