Now using Python 3 string formatting in rmgpy.chemkin module.
The idea here is to make things easier to switch to Python 3 once all
our dependencies are available on that platform. (I think we're just
waiting for matplotlib, maybe Django, and perhaps one or two others I
can't think of off the top of my head. Certainly numpy and scipy are
working, which is very important for us, as well as Cython and
openbabel.)