Moved rmgpy.chem.geometry module to rmgpy.cantherm.geometry.
The implementation of the Geometry class is somewhat specialized for use
in CanTherm, and isn't used anywhere else. Thus I am moving it in with
the CanTherm modules rather than the rmgpy package. Note that I have
deleted the Cython header file; this was only present so that we could
Cythonize the Species class. The functionality in the Geometry class is
likely not slow, and is not time-critical by any means, so we don't need
to Cythonize this module any more.

Also moved the corresponding unit tests and supporting files.