Refactoring of CanTherm to store geometries in new location.
With the removal of geometry as an attribute of the Species class, we
needed a new place to put this data during a CanTherm job. (We could
just not store it, but it's pretty nice to have in the output, e.g. if
you are putting together Supporting Materials for a manuscript.) It
turns out that this is easily done by creating a geometryDict in
rmgpy.cantherm.input to mirror the speciesDict, transitionStateDict, and
reactionDict module-level variables.

Note that the geometry is only required in CanTherm to compute the
reduced moment of inertia of the torsional modes, and to save to the
output file. (The geometry is needed to have energies and frequencies in
the first place, of course! However, this comes from the Gaussian output
already.)