Updated setup.py and Makefile to reflect new module locations.
The modules that were in rmgpy.chem are now mostly in rmgpy. (A few were
either placed in other locations or merged into other modules.) The new
setup.py script was updated to reflect this. If you want to compile just
the Cython modules that were in rmgpy.chem, use "main" instead of
"chem", i.e. ::

$ python setup.py build_ext main

or ::

$ make main

to use the Makefile.