Refactoring of rmgpy.cantherm modules to restore them to working order.
As is the theme with this branch, these edits fixed broken import
statements, converted Python 2 string formatting to Python 3 style, and
converted everything to work seamlessly with the Quantity object
framework. Testing on a few examples suggests that things seem to once
again be working. It would be nice to write some unit tests however!