Refactoring and cleanup of the rmgpy.measure modules.
Fixing broken import statements, correcting variables that are now
Quantity objects, switching to Python 3 string formatting syntax.
MEASURE seems to be back to working order, although there seems to be an
issue with the Chebyshev polynomial fitting method, in that it doesn't
give very good fits to the data. There was a lot of string formatting to
fix as well.