Refactoring and cleanup of the rmgpy.data modules.
Fixing broken import statements, correcting variables that are now
Quantity objects, switching to Python 3 string formatting syntax. I also
needed to add methods for adding ThermoData objects and multiplying
KineticsData objects, as I had earlier purposefully removed the __add__
and __mul__ methods of those classes since their functionality was not
generic enough to warrant operator overloading.

MEASURE seems to be back to working order, although there seems to be an
issue with the Chebyshev polynomial fitting method, in that it doesn't
give very good fits to the data. There was a lot of string formatting to
fix as well.