Created rmgpy.measure.main module for main MEASURE execution function.
The idea behind this change is to enable MEASURE jobs to also be invoked
in Python scripts, and not just via the measure.py script. Most of the
code in measure.py has been moved to rmgpy.measure.main. The main body
of a MEASURE job is now found in the execute() function.