Improved procedure for determining well order on drawn PES.
It turns out that this code was written before the
isomer/reactants/products attributes became part of the Network class.
As a result, it was making things far more complicated than necessary.
The actual order is simple: reactant channels first, isomers second, and
product channels last. This is done since people tend to read a PES from
left to right anyway.