In MEASURE input file, now must specify reactants and isomers explicitly.
Previously I had a pretty neat way of using the reaction reversibility
to determine this. However, the problem is that the the direction the
high-P k(T) is known may be the opposite of the direction needed for the
above procedure to work, especially for dissociation reactions. (This is
particularly true in RMG jobs, for which the kinetics is almost always
defined in the exothermic direction.)

In the new input file, users should provide all path reactions in the
direction for which kinetics is known. (This is only important if
using ILT to get the k(E) values; if using RRKM, then the reaction
direction does not matter.) To specify which configurations are isomers
and reactant channels, you will need to include additional code blocks
isomer() and reactants() for each. All other configurations will be
automatically set as product channels. Note that, if the input file has
any product channel-to-product channel reactions, an error will be
raised.

This change required revising the procedure by which path reactions are
checked to determine if they have the required thermo data. I've also
added a check to ensure that A + B -> C + D reactions aren't included in
the input file. This means that we don't need a special check for
available states data, since all isomers are required to have this
anyway, and each path reaction will involve at least one isomer.