Refactoring of calculateMicrocanonicalRateCoefficient() method.
The motivation here was to make this method easier to read/understand
and more robust to the types of input parameters it may be given. In
particular, the method needed to also handle the case where a reaction
of the form B + C -> A was given with B + C as a product channel and the
kinetics defined in the direction given. (This occurs fairly frequently
in RMG jobs, for example.)

The RRKM and ILT methods now have different criteria to determine
whether they are used to compute the forward or reverse k(E). The RRKM
method must be used in the direction for which the reactant is an isomer
(since it's a unimolecular theory). The ILT reaction will be used in the
direction the kinetics are given unless that direction involves a
reactant configuration without states data, in which case the reverse
direction will be used. (The ILT method is generally better used in the
exothermic direction due to the smaller Ea, so we prefer that if we have
the parameters to do it.)

In principle, this method should not have changed its return values
(unless there was a bug) except for the new functionality outlined
above. Some more testing is needed, though.