Changed input file syntax for specifying states data.
The old format was deficient due to poorly-named parameters (e.g. which
"symmetry" does the symmetry parameter refer to?) The new approach still
separates the types of degrees of freedom (rotations, vibrations,
torsions), but within each uses the classes from the rmgpy.statmech
module to simplify the implementation and encourage consistency across
the RMG-Py subprojects. This new syntax is cleaner, easier to
understand, easier to read and write, and easier to extend to new
statmech models.