Fixed units error in Chebyshev.fitToData() method.
The units error occurred when setting the minimum and maximum pressure
of the pressure range; the value was in Pa, but the units were given as
bar, leading to an error of 10^5 in the value. This bug was only
introduced recently, so it shouldn't have affected anything.

This bugfix addresses a warning that was popping up when using Chebyshev
fits for k(T,P) in MEASURE output.