Added kunits attribute to Chebyshev kinetics model.
A set of Chebyshev coefficients has an associated set of units, and it
is important to remember this to avoid units errors. To facilitate
this, I've added an attribute to the Chebyshev class for storing the
associated units, just to make it clear what they are. The unit tests
and various input/output routines have been updated to process this new
attribute.