Created new documentation sections for users, developers, and theory.
Before, the first categorization of the documentation was by component,
with the second being user's, developer's, or theory guide. The aim of
this commit is to reverse the order of categorization. This means all of
the user's guides will be together, all of the developer's guides will
be together, etc.