Added more informative error messages when loadOldDictionary() fails.
This is to make it easier for the user to find which adjacency list is
malformed by printing the file being read and the name of the adjacency
list that it stopped on. Note that logging.exception(), when used in an
except block, also seems to print the exception and its stack trace, so
I've switched to logging.error() for these messages.