Fixed bug in KineticsLibrary.loadOldReactions() method.
The way that this method determines the units of the rate coefficients
involves knowing the number of reactants. However, many of the queries
for units were doing so when the list of reactants incorrectly had a
length of zero. Thus, most A factors were not getting any units at all.
This fix corrects the error.