Kinetics depositories can now handle real species and functional groups.
The motivation here is so that we can use both the rate rules from the
old RMG-Java kinetics database and new data for reactions involving real
species in the training and/or test set for fitting group additivity
values. This way we can use all the rate rules we currently have to
train group values, plus add more as we collect them. Ideally, the rate
rules would eventually be deleted, but that need not happen for a while.