Now loading rate rules to kinetics depository when importing old database.
The purpose of the kinetics depository is to provide a place for us to
store the training and test data used to fit group additivity values.
This data can take the form of generic rate rules (such as those used in
the old kinetics rateLibrary.txt files) or kinetics for real reactions.
Eventually we want to use only the latter for training group values, but
for now we will allow both so we can train group values to the current
rate rules.

The other purpose is to enable exporting to the old database format
without loss of information. Since the kinetics groups are designed for
group additivity values only, we needed a place to put the rate rules
in the new format. The kinetics depository works well for this purpose.

Note that the kinetics depository can contain multiple sets of training
and test data, so we group them in folders by reaction family.