Moved handling of duplicate entries from parseOldLibrary to loadOldLibrary.
There are certain use cases where we want to keep duplicate entries
(e.g. the kinetics depository). For this reason, we use
parseOldLibrary() to load all of the library entries, including
duplicates, and leave it to the items that use this method to handle the
duplicate entries in the appropriate manner.

Since we don't want duplicate entries when loading an old database
normally (via loadOldLibrary), code to remove them (and log messages
about them) was moved to that method instead.