Added optional rank attribute to Entry class.
In the old RMG-Java kinetics rate rules, the rank is used to indicate
some degree of certainty in the rule. I don't think we really use this
metadata any more, but we might as well keep it around. Thus, the rank
attribute was added to the Entry class so that we can handle it if it
exists. So far I have only implemented saving it for the kinetics
database, as that's the only place we use it at the moment. Since I
expect that the rank will be removed at some point in the near future, I
don't want to provide general support for it.