Added duplicate attribute to Reaction objects.
This attribute is used e.g. in Chemkin files to indicate whether the
user wants to explicitly allow a reaction to exist multiple times in the
mechanism. If not set, an error would be raised if a duplicate reaction
was encountered. Adding the duplicate attribute to the Reaction object
enables the possibility of mimicking this functionality in RMG.