Reset connectivity values when switching between implicit/explicit hydrogens.
The connectivity values for the two cases are different -- this is why
you must have both molecules in the same mode when doing graph
isomorphism comparisons -- but they were not getting reset when
switching. This would potentially cause us to think that two
representations of the same molecule were different, and caused
duplicates to appear where not desired.