Now using Species instead of Molecule in reactions in kinetics libraries.
The motivation for this change is to allow for storing of a string label
with each species. The label is useful when paired with an adjacency
list because it allows for quick interpretation of what structure the
adjacency list represents. For example, you might label the
methylformate adjacency list as "Mfmt" as a clue to its structure.

One challenge with this approach is ensuring that each label corresponds
to a unique species. (This was easy in the old database, but is harder
now that we are designing each reaction entry to be standalone.) The new
getSpecies() method returns a dict of the unique species, and also
checks that each label does in fact represent the same species in all
uses.