Added KineticsLibrary.saveOld() method for exporting the old database.
In order to get people to switch to the new database, we need to be able
to export to the old database so that they can still use the new
values in RMG-Java until RMG-Py is ready. The saveOld() method provides
this functionality, which writes the species dictionary (species.txt)
and Chemkin-like reactions (reactions.txt and pdepreactions.txt) for the
kinetics library, including duplicate tags where appropriate.