Added functionality for saving old kinetics group database.
This is fairly straightforward, except that the rate rules are stored in
a kinetics depository, while the dictionary and tree are stored in a
kinetics groups database. Note that we have lost the original name of
the reverse direction for the reaction families that are not their own
reverse. Otherwise the information seems to be preserved intact.

We are also writing the comments.rst file using the longDesc attribute
of each entry.