Added forbidden structures database to RMGDatabase class.
The forbidden structures database is a place to collect functional
groups and molecules that RMG might naively create in the process of
mechanism generation, but are thought to not exist physically. Such a
database consists solely of a dictionary of structures. A
ForbiddenStructures class was created in the rmgpy.data.base module to
handle loading, saving, and querying of this database. A
forbiddenStructures attribute was added to the RMGDatabase class as a
place to store the forbidden structures in memory.