Added cache for unit conversion factors to Quantity class.
The quantities package, which we use to handle unit conversions, is not
at all optimized for speed. This is particularly noticeable when we have
a lot of unit conversions to process, e.g. when loading the database.
This commit adds a class variable conversionFactors that is used to
cache the conversion factors for various units as they are encountered,
so that quantities is only invoked once per units. This is a very small
cost in memory (as the number of units used is generally not very large)
but a significant speed boost.