Fixed type of degeneracy attribute of Reaction objects in Cython header.
The degeneracy is by definition an integer, but for some reason I had left
it as a double in the Cython header. This was discovered due to a string
formatting operation performed on the reaction object, which was correctly
expecting an integer but was receiving a double in compiled Cython mode.