Only load all kinetics libraries automatically if parameter is None.
In a recent commit I had also allowed for automatic loading of all
kinetics libraries when the `libraries` parameter is an empty list. This
has the unexpected side effect of causing all kinetics libraries to be
loaded in RMG jobs where no reaction libraries or seed mechanism are
specified. Although the database loads much faster due to the caching of
conversion factors in Quantity objects, this is still undesirable
behavior, so I have reverted it.