Restored cm^-1 as unit with custom conversion factor.
Much of the code (particularly the rmgpy.statmech module) is designed as if
cm^-1 is the characteristic unit of frequency. Before we had added a custom
conversion factor in order to make this work. In the refactoring and caching
of the conversion factors (to make them go faster) we inadvertently deleted
the overriding conversion factor for cm^-1. This has now been restored.