Added getAllSpecies() method to Network class.
This method provides an easy way to get a list containing all of the species
that participate in the pressure-dependent reaction network, including all
isomers, reactant and product channels, and even bath gas species. This is
already useful in output file writing; I expect it to be useful in other
situations as well.