Unified MEASURE input and output file formats.
As of now the only difference between the two is the presence of
pdepreaction() blocks in the latter, signifying the net reactions and their
pressure-dependent kinetics. This means that the entirety of the input file
is contained in the output file. Accordingly, there are no longer separate
readInput() and readOutput() methods, nor separate writeInput() and
writeOutput() methods; instead, there are now simply readFile() and
writeFile() methods to use for both cases.