HTML ouptput of RMG run contains reaction kinetics and comments.
The HTML page also has the ability to toggle display of these details.

The separation between formatting and function has been slightly blurred 
by this approach (putting a toHTML() method in the actual class definition)
but I think it offers the most flexibility, at least until we can think
of a "better" way.