Now fitting and saving k(T,P) values in both directions in MEASURE.
Previously we were only keeping the k(T,P) values in one direction, where
that direction was chosen solely based on the order of configurations in
the network. This is almost certainly not a good method of choice. Instead
of forcing one on everyone, however, I think it's better to leave it up
to the end user to decide what to do with things. Some applications might
prefer to keep both directions, while others might just want the
exothermic direction, or perhaps something else.