Fixed __reactionMatchesReactants() to work with Species objects.
Recently JWA had changed the kinetics database to use Species objects
instead of Molecule objects for the reactants and products of the
loaded reactions. This caused the __reactionMatchesReactants() method,
which assumed Molecule objects for the reactants and products, to
break. This commit fixes them to use Species objects, and adds a few
asserts to clearly note the assumptions of the method.

Additionally, in generateReactions() we commented out the searching
of the depository, since it can contain rate rules as well as real reactions,
and these would not work in __reactionMatchesReactants().