Bugfixes to the RMG HTML and Chemkin output methods.
In the HTML output, not all reactions have a family attribute; only the
TemplateReaction objects (that come from a reaction family) do. The proper
approach instead is to call the getSource() method, which will return the
family for a TemplateReaction and the library for a LibraryReaction.

In the Chemkin output, better handling of colliders in ThirdBody reactions
was needed, as these are currently represented as Molecule objects, when
Species objects are needed for the Chemkin output. I added some code to
determine the species that the collider corresponds to so we can use it
properly in the Chemkin file.