Updated the MEASURE documentation to reflect recent changes.
Most important are the changes to the input and output file syntax. Since
we probably won't have a web form or GUI for generating an input file for
some time, it is extremely important that this documentation be complete,
comprehensible, and accurate. Also important is the installation procedure,
which has changed somewhat since MEASURE was merged into the RMG-Py source
tree. However, I expect most users to start with the web interface, so
installation may be less important in the near term.

Also fixed the API documentation to reflect the new functions for reading
and writing MEASURE files.