Added errorString attribute to Network class.
The errorString attribute is used by the readFile() function to store the
error messages that are generated when an invalid MEASURE input file is
loaded. This enables them to remain available after readFile() returns,
so that they can be inspected or displayed in the future.

Note that readFile() now returns the loaded data even when the network
is invalid. Therefore, a proper test for a valid input file is that the
return value from readFile() is not None *and* the errorString of the
returned network is empty.