Moved detailed exception message to more efficient place.
The detailed exception message introduced a couple of commits ago was 
generated for every species that does not have gauche, 1-5, or 'other' 
corrections. A significant fraction of CPU time was spent formatting this
message which is never seen. I have moved it so only when it is an actual
error that would be uncaught is the detail generated and printed.