Added RMG class to rmgpy.rmg.main module for representing RMG jobs.
The idea is to gather the module-level variables and functions into a
single place, to better encapsulate functionality and thereby make things
easier to find and diagnose.

Lots of work toward fixing the issues caused (purposefully) by removing
the rmgpy.rmg.settings module has been done. Things should more or less
be working again, but there is definitely some room for improvement in
terms of organization of the code.