Wipe the output.html file from previous run by writing a blank one sooner.
As soon as the input file has been read in, we write an output file 
which has 0 species and 0 reactions. The species pictures etc. from the 
previous run have been wiped by this point already, but it can be confusing
to have the previous output file still there when you're running a new job.