Added duplicate parameter to __init__() of several Reaction derived classes.
Recently I added duplicate as a parameter to the Reaction base class that
DepositoryReaction, TemplateReaction, and LibraryReaction derive from. Since
I was not using keyword arguments, these derived classes were inadvertently
setting the degeneracy as the value of duplicate. I have added keyword
arguments to the call to the base class constructor so that this does not
happen in the future.