Now saving core species concentration profiles for each reaction system.
We again use the xlwt package to save the concentration profiles as
Excel spreadsheets, which lets us apply a bit of formatting to the data.
Each round of simulation gets its own spreadsheet; these are placed in
the new solver subdirectory. Each reaction system gets a sheet in the
spreadsheet for that iteration. Some information about the reaction
system is printed in the top few lines of each corresponding sheet.

I tried to design it so that RMG would carry on if xlwt was not available.
In the future we may want to make xlwt a required package.