Added KineticsFamily class for grouping database components by family.
Recently the organization of the kinetics database on disk was modified to
place the kinetics groups and kinetics depository for each reaction family
together under the "families" banner. This makes sense because these items
are closely related to each other. The objective now is to update the code
to reflect the same organization.

This commit provides the first version of a new KineticsFamily class that
implements the same logical grouping. Within this class, the groups and
depositories continue to use the KineticsGroups and KineticsDepository
classes. At the moment we explicitly indicate the database components
(groups, rules, training, test, (optional) PrIMe); this may not be the
best approach.

Only loading and saving of the kinetics families (from either the old or
new database formats) has been implemented at the moment.