Moved templates and recipes from KineticsGroups to KineticsFamily.
These items reflect the family as a whole, not just the implementation of
group additivity kinetics, so they belong in the KineticsFamily class. This
is still a bit awkward because the corresponding information is still
stored with the group additivity values in the database on disk, but it
works for now.