Bugfix in KineticsGroups.__generateReactions().
The reactants parameter to this method should be a list of lists of
Molecules (to allow for multiple resonance forms for each reactant
molecule). The parameter also handles a list of Molecules (by neglecting
resonance forms), but converting from this to the list of lists was
overwriting the original reactants parameter, which is bad form. Instead
a copy is made as this action is taken so that the original parameter
is not modified.