Fixed bug in fromAdjacencyList() function.
The bug involved converting the integer index of each atom in the adjacency
list to its corresponding Atom object. The for loop for doing this assumed
that the dictionary of keys would always be sorted, which is incorrect for
a dict object in Python. Thus, a pair of sort method calls were inserted to
ensure that the sorted criterion was met in all cases.