Fixed bug in KineticsDatabase.generateReactions() method.
The bug appears when you try to provide products to this method, as it
tries to call Molecule.isIsomorphic() with a Species object as the
parameter. This is not allowed, but Species.isIsomorphic() can be called
with a Molecule object as the parameter, so the answer is simply to swap
which is the parameter and which is the object.