Temporary bug fix for finding kinetics in R_Recombination family.
The rmgpy.data.kinetics.KineticsGroups instance 'R_Recombination'
is now called 'R_Recombination/groups', but there was a hard-coded check 
for the family name. For now, I have changed this to a .startswith() call,
so that it works again, but I wonder if we can remove this hard-coding?
It was introduced in 1ca7db26c0eb1615087f127a5268e8f27644fdeb.