Updated readFile() and writeFile() to use new MEASURE class.
Since a MEASURE job object gathers all of the parameters needed for a
pressure dependence calculation under one class, it is much cleaner to use
as part of the file reading and writing. For example, we don't need as
many module-level global variables in the input module as a result. We
also don't need to return a large tuple of parameters from readFile(),
or pass the same to writeFile(), which makes them more extensible in case
we need to add parameters in the future.

Part of this process involved changing many parameters to be stored as
Quantity objects instead of simple floats or lists. This makes the saved
files have the same units as the loaded files, which helps with
consistency.