Fixed a few bugs in PDepNetwork merging during RMG jobs.
There were two issues: (1) the lists of explored isomers were not getting
merged properly, and (2) the lists of product channels were not being checked
for isomers that the other network had explored. Both of these could result
an isomer mistakenly appearing as a product channel, leading to a path
reaction of the form reactant channel -> product channel, which is not
allowed and causes an exception to be raised.

Closes #41.