Handle case where productStructures is None when generating template reactions.
The __generateProductStrucutres() method can return None if applying the
reaction recipe fails to create the expected number of products. This could
be an error, but it also happens legitimately in certain cases
(particularly those involving intramolecular vs. intermolecular processes).
Thus the solution is to handle the case where productStructures is None,
as done in this commit.