Restored convertFAME.py script to working order.
This script was written before the major refactoring I did a few months
ago, so it had a lot of import statements that were out of date. It also
did not use Quantity objects for anything; introducing these makes the
script a bit more flexible, though it's still mostly geared to the FAME
input files generated during RMG-Java jobs.