Collision efficiency and matrix generation now accept dEdown as parameter.
This is done because the dEdown is often averaged among the two colliders
rather than using a single value for all isomer-bath gas collisions.

In the main Network.calculateRateCoefficients() method, the effective
dEdown for each isomer is computed separately so that it can be passed to
calculateCollisionEfficiency() and/or generateCollisionMatrix().