Added family attribute to DepositoryReaction and LibraryReaction.
This is needed for our reactionDict in RMG jobs, which organizes the
existing reactions to make for more rapid determinations if the generated
reactions are actually new or already existing. We can't use the result of
getSource() because it differs depending on how the kinetics were obtained,
when we don't really care about this distinction for reactionDict.