Fixed a few bugs in checkForExistingReaction().
Most importantly, the r1 and r2 variables were getting trampled over
within the for loop when checking for duplicate reactions in other seed
mechanisms because we were reusing them for checking the reverse direction
as well as the forward direction. Also fixed a reference to family in that
same loop that should have been family0.

These fixes decrease significantly the number of incorrectly-duplicated
reactions that appear in RMG output. We probably need to revisit actually
marked duplicate reactions at some point, though.