Better versions of incrementRadical() and decrementRadical() methods.
There was a nonobvious bug in the previous implementations that allowed a
spin multiplicity of zero, which is unphysical. In fixing this I rewrote
these methods to be a bit more clear about what they are doing.

This fixes a bug involving the reaction of allene with CH2(S), which
formed the species C(=C)([CH2])[CH2] which got stuck in an infinite loop
in generateResonanceIsomers().