Save restart file less frequently.
Since restart file saving is so slow, it is a waste of time to do it
after each and every iteration. Instead, a restart file is saved (1)
before the first ODE solve, (2) every hour or so after that, and (3)
at the end of the job. (If saveRestart=False in the input file, however,
the restart file is never saved.)