Don't hardcode kinetics family depositories.
These are now stored in a depositories dict on the KineticsFamily object.
The main reason for the previous hardcoding was so that we could
differentiate between the depositories of "recommended" data and those of
raw data. The new recommended flag means removes this barrier.

Note that the groups and rules remain hardcoded, as these are special
cases. (Eventually I expect to get rid of the rules and only have the
groups as a special case.)