Replace MultiArrhenius kinetics with more general MultiKinetics.
This represents kinetics that are meant to be a sum of other kinetics expressions.
Chemkin treats this as duplicate reactions. We treat it as a single reaction
with MultiKinetics. You can now have any combination of PDepArrhenius, 
Arrhenius, etc.

Also updated the unit tests, cython headers, and the databasa writing.
(It's a little annoying that the data/kinetics.py database writing
duplicates so much of the kineticse.py __repr__ code, but never mind for now)