PDepArrhenius class now has highPlimit attribute.
This stores an Arrhenius object which represents the high-P limit.
In Chemkin files this is usually printed on the line with the reaction.
It is not used in evaluating the k(T,P), but can store helpful
information regarding the source of the data anyway. Some of our reaction
libraries have this data, so we should have a place to store it.

Added test for PDepArrhenius highPlimit

Save highPlimit kinetics when reading old reaction libraries