Removed remaining references to old rmgpy.rmg.settings module.
There were a few places scattered around the code that referred to the
recently-removed rmgpy.rmg.settings module. I have replaced these with
references to the new locations that the settings are stored (i.e. in an
RMG object). This is still a bit hackish, and could use some more
refactoring to take advantage of the new approach.