Refactoring of RMG input file saving code to use RMG object.
This makes it so we can replace the InputFile class with the RMG class, so
I've also deleted the InputFile class. This commit also cleans up some
mixed tabs and spaces in the save input file code and converts it to use
spaces instead of tabs in the saved file itself.