Move KineticsDatabase.__reactionMatchesReactants to Reaction.matchesMolecules
This is essentially isomorphism checking, so now lives next to the real
isomorphism checking in the Reaction class. This also allows it to be
cythoned. I was hoping for a speed bost as a result, but it seems to be
pretty negligible - perhaps there's not enough static typing.

Certainly some pretty dumb hashing could probably speed us up a lot.
(eg. compare the molecular weight of both reactions?)
Or don't iterate through the whole library, but have them indexed by
one of the reactants (do we have Species objects at this point?).