New input file option: saveConcentrationProfiles
The xlwt package is taking a lot of time to save my concentration profiles
for the methylformate job, so it's nice to be able to turn it off
(unless we can find a faster method? CSV? or only save it occasionally?)
Anyway, there's now the option to turn it off.