ReactionFamily.generateReactions deals better with reactions that are their own reverse.
When asked to generate reactions of A+B, if A+B=C+D is created from a
family that is its own reverse, then the kinetics for C+D=A+B are also
estimated. If returnAllKinetics==True, then all possible estimates for
both directions are added to the list of returned reactions. (useful for
website).
If returnAllKinetics==False (useful for RMG runs) then only one of them
is returned. Currently the decision is to take the faster. Maybe this wants
to be looked at, but the important thing is that it's deterministic and
that only one is returned, to avoid duplicates.