When chosing reaction direction, cope with Evans Polanyi rates (assume dH=0)
If both forward and reverse directions match 'rules', which have EP rates,
then we assume dH=0 to estimate the rates then choose the faster direction.
Usually the EP alpha values are all zero so this usually won't introduce an 
error. The main thing is that it's deterministic. 

Perhaps a better way would be to postpone the direction decision until the 
enthlapy is known, but this should function for now.